.agents/skills/zenmux-image-generation/pyproject.toml
[project]
name = "zenmux-image-generation"
version = "0.1.0"
description = "ZenMux image generation helper scripts for the zenmux-image-generation skill."
requires-python = ">=3.14.5"
dependencies = [
    "google-genai>=1.73.1",
    "httpx[socks]>=0.28.1",
    "openai>=2.36.0",
]

[tool.uv]
package = false
